requirements.txt
# =============================================================================
# AUTOMATED BOOK GENERATION SYSTEM - DEPENDENCIES
# =============================================================================

# Core dependencies
python-dotenv>=1.0.0

# LLM Providers (install the one you're using)
openai>=1.0.0
anthropic>=0.20.0
google-generativeai>=0.5.0

# Web UI (optional)
flask>=3.0.0

# Database
# SQLite is built-in to Python (default, no install needed)
# supabase>=2.0.0  # Optional: uncomment for Supabase (requires C++ Build Tools)

# File handling
pandas>=2.0.0
openpyxl>=3.1.0  # For Excel files

# Document export
python-docx>=1.0.0  # For DOCX export
reportlab>=4.0.0    # For PDF export

# Notifications
requests>=2.31.0  # For MS Teams webhooks

# Web search (optional)
google-search-results>=2.4.0  # SerpAPI for web search

# Development dependencies (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0
